requirements.txt
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
DateTime==5.5
google-api-core==2.20.0
google-auth==2.35.0
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
idna==3.10
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
urllib3==2.2.3
zope.interface==7.0.3
